Manage categories

+++ app/Http/Controllers/AdminController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Category;
+use App\Product;
 
 class AdminController extends Controller
 {
@@ -27,7 +28,8 @@ class AdminController extends Controller
 			if($id !== '_token'){
 				if(($data['type'] == 'remove') && ($id >= 1)){
 					$found = Category::findOrFail($id);
-					$found->delete();
+					if($found->products->isEmpty())
+						$found->delete();
 				}
 				else if(($data['name'] !== '') && ($data['type'] !== '')){
 					if($id >= 1)
@@ -42,5 +44,19 @@ class AdminController extends Controller
 		}
 		
 		return redirect()->route('categories.index');
+	}
+	
+	public function showProducts()
+	{
+		$products = Product::all();
+		$food = Category::food()->get();
+		$drinks = Category::drinks()->get();
+		$other = Category::other()->get();
+		return view('admin.products', compact('products', 'food', 'drinks', 'other'));
+	}
+	
+	public function updateProducts()
+	{
+		return redirect()->route('products.index');
 	}
 }
